refactor(virtual-environment): change from virtual env (venv) to pipenv
BREAKING CHANGE: pipenv makes it easier(automatic) the update process of the packages to the latest version. Allow us to install packages only for dev and other for prod. Save the hashes of each package and sub-dependency, etc, etc. Also the package commitizen was installed to set a standard in the commits from the beginning of the project.